Bug 825949 followup: in nsGlobalWindow::SetTimeoutOrInterval, make |realInterval| unsigned, since it's known-nonnegative and is passed elsewhere as an unsigned integer. r=smaug
UltraBlame original commit: 8b397adc49e9c280d469821402c972aaeca45fa2

+++ dom/base/nsGlobalWindow.cpp
@@ -9571,11 +9571,11 @@ nsGlobalWindow::SetTimeoutOrInterval(nsIScriptTimeoutHandler *aHandler,
 
   
   uint32_t nestingLevel = sNestingLevel + 1;
-  int32_t realInterval = interval;
+  uint32_t realInterval = interval;
   if (aIsInterval || nestingLevel >= DOM_CLAMP_TIMEOUT_NESTING_LEVEL) {
     
     
-    realInterval = NS_MAX(realInterval, DOMMinTimeoutValue());
+    realInterval = NS_MAX(realInterval, uint32_t(DOMMinTimeoutValue()));
   }
